backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
google-generativeai>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
PyPDF2>=3.0.0
python-multipart>=0.0.6
redis>=4.5.0
python-dotenv>=1.0.0
slowapi>=0.1.7
numpy>=1.24.0